+++ src/pdoc-commons/model/records/pdoc-record.ts
@@ -5,9 +5,12 @@ export class PDocRecord extends BaseEntityRecord {
     descTxt: string;
     descMd: string;
     descHtml: string;
-    flgShowSearch: boolean;
-    flgShowNews: boolean;
-    flgShowTopTen: boolean;
+    flgShowSearch?: boolean;
+    flgShowNews?: boolean;
+    flgShowTopTen?: boolean;
+    flgShowAdminArea?: boolean;
+    flgShowDashboard?: boolean;
+    flags?: string[];
     heading: string;
     image: string;
     keywords: string;